requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
deepspeed>=0.9.0

# TerraTorch dependencies
pyyaml>=6.0
pytorch-lightning>=2.0.0
timm>=0.9.0
einops>=0.6.0

# Utilities
pandas>=2.0.0
matplotlib>=3.7.0
tabulate>=0.9.0
tqdm>=4.65.0
scikit-learn>=1.2.0

# Optional: for large model optimization
triton>=2.0.0
flash-attn>=2.0.0

# Optional: for geospatial data processing
rasterio>=1.3.0
geopandas>=0.12.0
